+++ heftia-effects/src/Control/Monad/Hefty/Concurrent/Timer.hs
@@ -1,10 +1,12 @@
 -- SPDX-License-Identifier: MPL-2.0
 
-module Control.Monad.Hefty.Concurrent.Timer where
+module Control.Monad.Hefty.Concurrent.Timer (
+    module Control.Monad.Hefty.Concurrent.Timer,
+    module Data.Effect.Concurrent.Timer,
+)
+where
 
 import Control.Concurrent.Thread.Delay qualified as Thread
-import Control.Monad.Hefty.Coroutine (runCoroutine)
-import Control.Monad.Hefty.State (evalState)
 import Control.Monad.Hefty (
     interpose,
     interpret,
@@ -17,7 +19,9 @@ import Control.Monad.Hefty (
     type (<|),
     type (~>),
  )
-import Data.Effect.Concurrent.Timer (CyclicTimer (Wait), Timer (..), clock, cyclicTimer)
+import Control.Monad.Hefty.Coroutine (runCoroutine)
+import Control.Monad.Hefty.State (evalState)
+import Data.Effect.Concurrent.Timer
 import Data.Effect.Coroutine (Status (Continue, Done))
 import Data.Effect.State (get, put)
 import Data.Time (DiffTime)